+++ .github/workflows/publish-openapi.yaml
@@ -10,5 +10,5 @@ jobs:
     secrets:
       SWAGGER_PUBLISHER_API_TOKEN: ${{ secrets.SWAGGER_PUBLISHER_API_TOKEN }}
     with:
-      test_to_run: 'uk.gov.hmcts.reform.demo.openapi.OpenAPIPublisherTest'
+      test_to_run: 'uk.gov.hmcts.reform.laubackend.eud.openapi.OpenAPIPublisherTest'
       java_version: 17
