+++ Dockerfile
@@ -3,7 +3,7 @@ ARG APP_INSIGHTS_AGENT_VERSION=3.7.3
 FROM hmctspublic.azurecr.io/base/java:21-distroless
 
 COPY lib/applicationinsights.json /opt/app/
-COPY build/libs/spring-boot-template.jar /opt/app/
+COPY build/libs/lau-eud-backend.jar /opt/app/
 
 EXPOSE 4550
-CMD [ "spring-boot-template.jar" ]
+CMD [ "lau-eud-backend.jar" ]
