+++ Jenkinsfile_CNP
@@ -5,7 +5,7 @@
 @Library("Infrastructure")
 
 def type = "java"
-def product = "rpe"
-def component = "demo"
+def product = "lau"
+def component = "eud-backend"
 
 withPipeline(type, product, component) {}
