+++ Jenkinsfile_nightly
@@ -8,7 +8,7 @@ properties([
 @Library("Infrastructure")
 
 def type = "java"
-def product = "rpe"
-def component = "spring-boot-template"
+def product = "lau"
+def component = "eud-backend"
 
 withNightlyPipeline(type, product, component) {}
