+++ build.gradle
@@ -13,7 +13,7 @@ plugins {
 }
 
 application {
-  mainClass = 'uk.gov.hmcts.reform.demo.Application'
+  mainClass = 'uk.gov.hmcts.reform.laubackend.eud.Application'
   group = 'uk.gov.hmcts.reform'
   version = '0.0.1'
 }
@@ -102,8 +102,8 @@ tasks.check.dependsOn(integration)
 
 sonarqube {
   properties {
-    property "sonar.projectName", "Reform :: spring-boot-template"
-    property "sonar.projectKey", "uk.gov.hmcts.reform:spring-boot-template"
+    property "sonar.projectName", "Reform :: lau-eud-backend"
+    property "sonar.projectKey", "uk.gov.hmcts.reform:lau-eud-backend"
   }
 }
 
@@ -158,7 +158,7 @@ dependencies {
 }
 
 bootJar {
-  archiveFileName = "spring-boot-template.jar"
+  archiveFileName = "lau-eud-backend.jar"
 
   manifest {
     attributes('Implementation-Version': project.version.toString())
