+++ charts/lau-eud-backend/Chart.yaml
@@ -1,0 +1,12 @@
+apiVersion: v2
+appVersion: "1.0"
+description: A Helm chart for lau-eud-backend App
+name: lau-eud-backend
+home: https://github.com/hmcts/lau-eud-backend
+version: 0.0.17
+maintainers:
+  - name: HMCTS lau team
+dependencies:
+  - name: java
+    version: 5.3.0
+    repository: 'oci://hmctspublic.azurecr.io/helm'
