--- charts/rpe-spring-boot-template/Chart.yaml
@@ -1,12 +1,0 @@
-apiVersion: v2
-appVersion: "1.0"
-description: A Helm chart for spring-boot-template App
-name: rpe-spring-boot-template
-home: https://github.com/hmcts/spring-boot-template
-version: 0.0.17
-maintainers:
-  - name: HMCTS rpe team
-dependencies:
-  - name: java
-    version: 5.2.1
-    repository: 'https://hmctspublic.azurecr.io/helm/v1/repo/'
