+++ lib/applicationinsights.json
@@ -1,7 +1,7 @@
 {
-  "connectionString": "${file:/mnt/secrets/rpe/app-insights-connection-string}",
+  "connectionString": "${file:/mnt/secrets/lau/app-insights-connection-string}",
   "role": {
-    "name": "rpe-demo"
+    "name": "lau-eud-backend"
   },
   "sampling": {
     "overrides": [
